abstract contact & room
